backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
python-dotenv
pydantic
httpx
Pillow
numpy
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
loguru
google-generativeai
psutil
